[Feat] Character Set Dead
캐릭터(Monster) 사망 처리

+++ Source/RLR/UI/InGame/CharacterDisplay/MonsterStatDisplay.h
@@ -34,7 +34,7 @@ public:
 
 protected:
 	bool bCompletedChange;
-	int MaxHp;
-	int CurrentHp;
-	int TargetHp;
+	float MaxHp;
+	float CurrentHp;
+	float TargetHp;
 };
